% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCValues.R
\name{getCValues}
\alias{getCValues}
\title{Get C Value}
\usage{
getCValues(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a numeric vector of the C value for each scan
}
\description{
Get the FT-MS conversion parameter \code{C} for each scan in a given range
}
\examples{
\dontrun{
getCValues(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

