% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getITtime.R
\name{getITtime}
\alias{getITtime}
\title{Get Ion Injection Time}
\usage{
getITtime(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a numeric vector of the ion injection time for each scan
}
\description{
Get the ion injection time (IT) for each scan in a given range
}
\examples{
\dontrun{
getITtime(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

