% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMzIntNoise.R
\name{getMzIntNoise}
\alias{getMzIntNoise}
\title{Get Mass, Intensity and Noise}
\usage{
getMzIntNoise(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a list of matrices for m/z, intensity and noise
}
\description{
Get the profile data (m/z, intensity and noise) across a given scan range
}
\examples{
\dontrun{
getMzIntNoise(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

