% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResComp.R
\name{getResComp}
\alias{getResComp}
\title{Get Resolution Compensation}
\usage{
getResComp(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a numeric vector of the Resolution compensation for each scan
}
\description{
Get the Resolution mass compensation (ppm) for each scan in a given range
}
\examples{
\dontrun{
getResComp(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

