% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRFComp.R
\name{getRFComp}
\alias{getRFComp}
\title{Get RF Compensation}
\usage{
getRFComp(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a numeric vector of the RF compensation for each scan
}
\description{
Get the RF mass compensation (ppm) for each scan in a given range
}
\examples{
\dontrun{
getRFComp(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

