% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpaceComp.R
\name{getSpaceComp}
\alias{getSpaceComp}
\title{Get Space Charge Compensation}
\usage{
getSpaceComp(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a numeric vector of the space charge compensation for each scan
}
\description{
Get the space charge mass compensation (ppm) for each scan in a given range
}
\examples{
\dontrun{
getSpaceComp(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

