% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFilter.R
\name{scanFilter}
\alias{scanFilter}
\title{Get Filter}
\usage{
scanFilter(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a character vector of the scan filter for each scan
}
\description{
Get the scan filter for each scan in a given range
}
\examples{
\dontrun{
scanFilter(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

