% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPackage.R
\name{testPackage}
\alias{testPackage}
\title{Package Testing}
\usage{
testPackage(scans = c(1:5))
}
\arguments{
\item{scans}{a numeric vector of scan numbers to extract}
}
\description{
Run some basic tests to ensure there have been no problems during compilation
}
\examples{
\dontrun{
	testPackage(scans = c(1:5))
}
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

