% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMzInt.R
\name{getMzInt}
\alias{getMzInt}
\title{Get Mass and Intensity}
\usage{
getMzInt(filename, scans = c())
}
\arguments{
\item{filename}{a \code{.raw} file}

\item{scans}{a numeric vector of scan numbers to extract}
}
\value{
a list of matrices for m/z and intensity.
}
\description{
Get the profile data (m/z and intensity) across a given scan range
}
\examples{
\dontrun{
getMzInt(QC.raw, scans = c(2:24)	
}

}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

